Refactor school vs general page structure into separate folders
