+++ src/app/(general)/founder-archetypes/page.tsx
