+++ src/app/(general)/pricing/page.tsx
