+++ src/app/(school)/layout.tsx
@@ -1,0 +1,7 @@
+export default function SchoolLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <>{children}</>;
+}
