+++ src/app/(school)/school/[slug]/admin/layout.tsx
