+++ src/app/(school)/school/[slug]/admin/page.tsx
