+++ src/app/(school)/school/[slug]/dashboard/page.tsx
