+++ src/app/(school)/school/[slug]/layout.tsx
