+++ src/app/(school)/school/[slug]/page.tsx
