+++ src/app/(school)/school/[slug]/pending-activation/page.tsx
