+++ src/app/(school)/school/[slug]/profile/page.tsx
