+++ src/app/(school)/school/[slug]/sign-up/[[...sign-up]]/page.tsx
