+++ src/app/(school)/school/[slug]/sso-callback/page.tsx
