+++ src/app/(school)/school/[slug]/sso-complete/page.tsx
